Remove static methods
